Rename component names
SupportedCamps=>DirectSupportedCamps
